YamlConfig -> Config + cmd args for a config file
