+++ src/main/java/Game/Monster.java
@@ -1,7 +1,6 @@
 package Game;
 
 import Main.Config;
-import Main.YamlConfig;
 import lombok.Getter;
 import java.util.ArrayList;
 import java.util.List;
@@ -36,7 +35,7 @@ public class Monster extends GameCreature {
 
     public void makeDecision(short[] boardData) throws Exception {
 
-        YamlConfig config = YamlConfig.getInstance();
+        Config config = Config.getInstance();
         Action decidedAction = ((boolean) config.getConfig("deterministic_monsters")) ? makeDeterministicDecision() : makeRandomDecision(boardData);
 
         if (!Board.getState().checkActionValidity(currentPosition, decidedAction))
